requirements.txt
# Adversarial Defense Mechanisms - Requirements

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scipy>=1.7.0

# Data handling
pandas>=1.3.0
pillow>=8.3.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web UI
streamlit>=1.25.0

# Database
sqlite3  # Built-in with Python

# Configuration
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Optional: GPU acceleration
# torch-audio>=2.0.0  # Uncomment if needed
# torchaudio>=2.0.0   # Uncomment if needed
